--- api/src/main/java/com/beanbank/api/BeanBankApiApplication.java
@@ -1,11 +1,0 @@
-package com.beanbank.api;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BeanBankApiApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(BeanBankApiApplication.class, args);
-	}
-}
