--- api/src/main/java/com/beanbank/api/model/User.java
@@ -1,54 +1,0 @@
-package com.beanbank.api.model;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name="User")
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="UserID")
-    private int userID;
-
-    @Column(name="Username", unique=true)
-    private String username;
-
-    @Column(name="Password")
-    private String password;
-
-    @Column(name="Email", unique=true)
-    private String email;
-
-
-    public int getUserID() {
-        return userID;
-    }
-
-    public void setUserID(int userID) {
-        this.userID = userID;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-}
