--- api/src/main/java/com/beanbank/api/repository/AccountRepository.java
@@ -1,9 +1,0 @@
-package com.beanbank.api.repository;
-import com.beanbank.api.model.Account;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AccountRepository extends JpaRepository<Account, Integer> {
-
-}
