--- api/src/main/java/com/beanbank/api/repository/BeanTypeRepository.java
@@ -1,7 +1,0 @@
-package com.beanbank.api.repository;
-
-import com.beanbank.api.model.BeanType;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BeanTypeRepository extends JpaRepository<BeanType, Integer> {
-}
