--- api/src/main/java/com/beanbank/api/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package com.beanbank.api.repository;
-
-import com.beanbank.api.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Integer> {
-    User findByUsername(String username);
-}
