--- api/src/main/java/com/beanbank/api/service/TransactionTypeService.java
@@ -1,25 +1,0 @@
-package com.beanbank.api.service;
-
-import com.beanbank.api.model.TransactionsType;
-import com.beanbank.api.repository.TransactionsTypeRepository;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class TransactionTypeService {
-    final
-    TransactionsTypeRepository transactionTypeRepository;
-
-    public TransactionTypeService(TransactionsTypeRepository TransactionTypeRepository) {
-        this.transactionTypeRepository = TransactionTypeRepository;
-    }
-
-    public TransactionsType getTransactionTypeByID(int transactionTypeID){
-        return transactionTypeRepository.findById(transactionTypeID).get();
-    }
-
-    public List<TransactionsType> getTransactionTypes(){
-        return transactionTypeRepository.findAll();
-    }
-}
