+++ app/api/src/main/java/com/beanbank/api/BeanBankApiApplication.java
@@ -1,0 +1,17 @@
+package com.beanbank.api;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@EnableJpaRepositories("com.*")
+@ComponentScan(basePackages = { "com.*" })
+@EntityScan("com.*")
+@SpringBootApplication
+public class BeanBankApiApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(BeanBankApiApplication.class, args);
+	}
+}
