+++ app/api/src/main/java/com/beanbank/api/repository/AccountRepository.java
@@ -1,0 +1,10 @@
+package com.beanbank.api.repository;
+
+import com.models.Account;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AccountRepository extends JpaRepository<Account, Integer> {
+
+}
