+++ app/api/src/main/java/com/beanbank/api/repository/BeanTypeRepository.java
@@ -1,0 +1,7 @@
+package com.beanbank.api.repository;
+
+import com.models.BeanType;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BeanTypeRepository extends JpaRepository<BeanType, Integer> {
+}
