+++ app/api/src/main/java/com/beanbank/api/repository/TransactionsTypeRepository.java
@@ -1,0 +1,9 @@
+package com.beanbank.api.repository;
+import com.models.TransactionsType;
+import org.springframework.stereotype.Repository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+@Repository
+public interface TransactionsTypeRepository extends JpaRepository<TransactionsType, Integer> {
+
+}
