+++ app/shared/src/main/java/com/models/Transaction.java
@@ -1,0 +1,83 @@
+package com.models;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+
+@Entity
+@Table(name = "Transactions")
+public class Transaction {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "TransactionID")
+    private int transactionID;
+
+    @Column(name = "SenderID")
+    private int senderID;
+
+    @Column(name = "ReceiverID")
+    private int receiverID;
+
+    @Column(name = "TransactionTypeID")
+    private int transactionTypeID;
+
+    @Column(name = "TransactionAmount")
+    private BigDecimal transactionAmount;
+
+    @Column(name = "TransactionTimestamp")
+    private LocalDateTime transactionTimestamp;
+
+    public int getTransactionID() {
+        return this.transactionID;
+    }
+
+    public void setTransactionID(int transactionID) {
+        this.transactionID = transactionID;
+    }
+
+    public int getSenderID() {
+        return this.senderID;
+    }
+
+    public void setSenderID(int senderID) {
+        this.senderID = senderID;
+    }
+
+    public int getReceiverID() {
+        return this.receiverID;
+    }
+
+    public void setReceiverID(int receiverID) {
+        this.receiverID = receiverID;
+    }
+
+    public int getTransactionTypeID() {
+        return this.transactionTypeID;
+    }
+
+    public void setTransactionTypeID(int transactionTypeID) {
+        this.transactionTypeID = transactionTypeID;
+    }
+
+    public BigDecimal getTransactionAmount() {
+        return this.transactionAmount;
+    }
+
+    public void setTransactionAmount(BigDecimal transactionAmount) {
+        this.transactionAmount = transactionAmount;
+    }
+
+    public LocalDateTime getTransactionTimestamp() {
+        return this.transactionTimestamp;
+    }
+
+    public void setTransactionTimestamp(LocalDateTime transactionTimestamp) {
+        this.transactionTimestamp = transactionTimestamp;
+    }
+}
